refactor: extract messaging package
