+++ chrome-extension/lib/background/messageHandler.ts
@@ -1,4 +1,4 @@
-import { messaging } from '@extension/shared';
+import { messaging } from '@extension/messaging';
 
 export function addMessageHandler() {
   messaging.on('foo', async ({ bar }) => `Hello, ${bar}!`);
