+++ package.json
@@ -9,13 +9,13 @@
   },
   "scripts": {
     "clean": "turbo clean && turbo daemon stop && rimraf dist && rimraf .turbo",
-    "build": "turbo build",
+    "build": "turbo ready:once && turbo build",
     "build:firefox": "cross-env __FIREFOX__=true turbo build",
     "zip": "turbo zip",
     "zip:firefox": "cross-env __FIREFOX__=true turbo zip",
-    "dev-server": "pnpm -F hmr ready && pnpm -F hmr dev",
-    "dev": "turbo ready && turbo watch dev --concurrency 20",
-    "dev:firefox": "turbo ready && cross-env __FIREFOX__=true turbo watch dev --concurrency 20",
+    "dev-server": "pnpm -F hmr ready:once && pnpm -F hmr dev",
+    "dev": "turbo ready:once && turbo ready && turbo watch dev --concurrency 20",
+    "dev:firefox": "turbo ready:once && turbo ready && cross-env __FIREFOX__=true turbo watch dev --concurrency 20",
     "test": "turbo test",
     "type-check": "turbo type-check",
     "lint": "turbo lint --continue -- --fix --cache --cache-location node_modules/.cache/.eslintcache",
