+++ packages/hmr/package.json
@@ -14,7 +14,7 @@
     "clean": "rimraf ./dist && rimraf ./build",
     "build:tsc": "tsc -b tsconfig.build.json",
     "build:rollup": "rollup --config rollup.config.mjs",
-    "ready": "pnpm run build:tsc && pnpm run build:rollup",
+    "ready:once": "pnpm run build:tsc && pnpm run build:rollup",
     "dev": "node dist/lib/initializers/initReloadServer.js",
     "lint": "eslint . --ext .ts,.tsx",
     "lint:fix": "pnpm lint --fix",
