+++ packages/messaging/build.mjs
@@ -1,0 +1,15 @@
+import * as esbuild from 'esbuild';
+
+/**
+ * @type { import("esbuild").BuildOptions }
+ */
+const buildOptions = {
+  entryPoints: ['./index.ts', './messaging.ts'],
+  tsconfig: './tsconfig.json',
+  bundle: false,
+  target: 'es6',
+  outdir: './dist',
+  sourcemap: true,
+};
+
+await esbuild.build(buildOptions);
