+++ packages/messaging/index.ts
@@ -1,2 +1,1 @@
-export * from './type';
 export * from './messaging';
