+++ packages/shared/index.ts
@@ -1,4 +1,3 @@
 export * from './lib/hooks';
 export * from './lib/hoc';
-export * from './lib/message';
 export * from './lib/utils';
