+++ pages/content-ui/package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "@extension/shared": "workspace:*",
     "@extension/storage": "workspace:*",
+    "@extension/messaging": "workspace:*",
     "@extension/ui": "workspace:*"
   },
   "devDependencies": {
